dual_cross_attention/utils/attention_rollout.py
"""
Attention Rollout Implementation for Dual Cross-Attention Learning

This module implements attention rollout computation as described in the paper
and based on the reference implementation in vit_rollout.py. Attention rollout
tracks how information flows from input tokens through transformer layers by
recursively computing attention across all previous layers.

Key equation from paper:
Ŝ_i = S̄_i ⊗ S̄_{i-1} ⊗ ... ⊗ S̄_1
where S̄_l = 0.5S_l + 0.5E (accounts for residual connections)

This is used by GLCA to identify high-response local regions for cross-attention.
"""

import torch
import torch.nn as nn
import cv2
import numpy as np
from typing import List, Tuple, Optional, Dict
import matplotlib.pyplot as plt
from PIL import Image


class AttentionRollout:
    """
    Attention rollout computation for identifying important image patches.
    
    Based on the attention rollout method and vit_rollout.py implementation.
    Computes accumulated attention scores across all transformer layers to identify
    which input patches are most important for the model's predictions.
    
    This class is used by GlobalLocalCrossAttention to select high-response
    local regions (top R% patches) for discriminative feature learning.
    
    Args:
        discard_ratio: Ratio of lowest attention paths to discard (default: 0.9)
        head_fusion: Method to fuse multi-head attention ("mean", "max", "min")
        
    Reference: vit_rollout.py rollout() function
    """
    
    def __init__(self, discard_ratio: float = 0.9, head_fusion: str = "mean"):
        pass
    
    def rollout(self, attentions: List[torch.Tensor]) -> torch.Tensor:
        """
        Compute attention rollout across all transformer layers.
        
        Implements the core rollout algorithm from vit_rollout.py:
        1. Fuse attention heads using specified method
        2. Add identity matrix (0.5 * attention + 0.5 * identity) for residual connections
        3. Normalize attention weights
        4. Multiply attention matrices across all layers
        5. Extract CLS token attention to all patches
        
        Args:
            attentions: List of attention matrices from all SA layers
                       Each tensor shape: [batch_size, num_heads, seq_len, seq_len]
                       
        Returns:
            rollout_scores: CLS token attention scores for all patches
                          Shape: [batch_size, num_patches]
        """
        pass
    
    def fuse_attention_heads(self, attention: torch.Tensor) -> torch.Tensor:
        """
        Fuse multi-head attention using specified method.
        
        Args:
            attention: Multi-head attention tensor [batch, heads, seq, seq]
            
        Returns:
            fused_attention: Single attention matrix [batch, seq, seq]
        """
        pass
    
    def get_cls_attention_map(self, rollout_result: torch.Tensor, 
                             img_size: Tuple[int, int], patch_size: int = 16) -> np.ndarray:
        """
        Convert CLS attention scores to spatial attention map.
        
        Reshapes the 1D attention scores back to 2D spatial map matching
        the original image patch layout for visualization.
        
        Args:
            rollout_result: CLS attention scores [batch_size, num_patches]
            img_size: Original image size (height, width)  
            patch_size: Size of each patch (default: 16)
            
        Returns:
            attention_map: 2D spatial attention map [height/patch_size, width/patch_size]
        """
        pass


def rollout_attention_maps(model: nn.Module, input_tensor: torch.Tensor,
                          discard_ratio: float = 0.9, head_fusion: str = "mean") -> Dict[str, torch.Tensor]:
    """
    Compute attention rollout maps for a given model and input.
    
    Extracts attention weights from all transformer layers in the model
    and computes rollout maps for both SA and GLCA branches if available.
    
    Args:
        model: Dual cross-attention model
        input_tensor: Input image tensor [batch_size, channels, height, width]
        discard_ratio: Ratio of attention paths to discard
        head_fusion: Multi-head fusion method
        
    Returns:
        rollout_maps: Dictionary containing rollout maps for different branches
                     Keys: "sa_rollout", "glca_rollout"
    """
    
    # Hook function to collect attention weights
    attention_weights = {"sa": [], "glca": []}
    
    def get_attention_hook(name):
        def hook(module, input, output):
            # Extract attention weights from module output
            if hasattr(output, 'attention_weights'):
                attention_weights[name].append(output.attention_weights)
        return hook
    
    pass


def visualize_attention_rollout(image: np.ndarray, attention_map: np.ndarray,
                               save_path: Optional[str] = None, alpha: float = 0.6) -> np.ndarray:
    """
    Visualize attention rollout as overlay on original image.
    
    Based on show_mask_on_image() function from vit_rollout.py.
    Creates heatmap visualization of attention rollout overlaid on input image.
    
    Args:
        image: Original input image [height, width, 3]
        attention_map: Attention rollout map [height, width]
        save_path: Optional path to save visualization
        alpha: Blending factor for overlay (0.0 = only heatmap, 1.0 = only image)
        
    Returns:
        visualization: Combined image with attention overlay [height, width, 3]
    """
    pass


def extract_high_response_regions(rollout_scores: torch.Tensor, top_k_ratio: float = 0.1) -> torch.Tensor:
    """
    Extract indices of high-response regions from attention rollout.
    
    Used by GLCA to select top R% patches with highest attention scores
    for local discriminative feature learning. R=10% for FGVC, R=30% for Re-ID.
    
    Args:
        rollout_scores: CLS attention scores [batch_size, num_patches]
        top_k_ratio: Ratio of top patches to select (R in the paper)
        
    Returns:
        high_response_indices: Indices of selected patches [batch_size, num_selected]
    """
    pass


def compute_attention_statistics(attention_weights: List[torch.Tensor]) -> Dict[str, float]:
    """
    Compute statistics of attention patterns across layers.
    
    Provides insights into attention distribution and can help with
    model analysis and debugging.
    
    Args:
        attention_weights: List of attention matrices from all layers
        
    Returns:
        stats: Dictionary with attention statistics (entropy, max values, etc.)
    """
    pass


class AttentionHook:
    """
    Hook class for collecting attention weights during forward pass.
    
    Utility class to register hooks on attention modules and collect
    attention weights for rollout computation and visualization.
    """
    
    def __init__(self):
        pass
    
    def register_hooks(self, model: nn.Module, layer_names: List[str]):
        """
        Register hooks on specified attention layers.
        
        Args:
            model: Model to register hooks on
            layer_names: Names of layers to hook into
        """
        pass
    
    def get_attention_weights(self) -> Dict[str, List[torch.Tensor]]:
        """
        Get collected attention weights from all hooked layers.
        
        Returns:
            attention_weights: Dictionary mapping layer names to attention weights
        """
        pass
    
    def clear_hooks(self):
        """Remove all registered hooks"""
        pass

