requirements.txt
# Core deep learning framework
torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.12

# Data processing and utilities  
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.3.0
opencv-python>=4.5.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.62.0
argparse
dataclasses; python_version<"3.7"

# Logging and experiment tracking
wandb>=0.12.0
tensorboard>=2.8.0

# Image processing and augmentation
albumentations>=1.1.0
imgaug>=0.4.0

# Optional: for advanced visualizations
jupyter>=1.0.0
ipywidgets>=7.6.0

# Optional: for distributed training
accelerate>=0.12.0

# Development and testing (optional)
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0

# System utilities
psutil>=5.8.0

